Separated the periodic refresh provider to ensure the file only exports components...
