+++ frontend/src/components/ExchangeCourses/Pickup/PickupsList.jsx
@@ -4,7 +4,7 @@ import PickupCourseItem from "./PickupCourseItem";
 import Typography from "@mui/material/Typography";
 import { useApi } from "../../../contexts/ApiProvider";
 import { useRefresh } from "../../../contexts/RefreshProvider";
-import { usePeriodicRefresh } from "../../../contexts/PeriodicRefreshProvider";
+import { usePeriodicRefresh } from "../../../contexts/usePeriodicRefresh";
 
 const PickupsList = () => {
   const [courses, setCourses] = useState([]);
