+++ frontend/src/contexts/PeriodicRefreshContext.jsx
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+const PeriodicRefreshContext = createContext({
+  subscribe: () => {},
+  unsubscribe: () => {}
+});
+
+export default PeriodicRefreshContext;
